0.7.8 -solucion crasheo por version de graddle -Agregada informacion de seguidores y seguidos

+++ app/src/main/java/com/example/chris/sipho/AdaptadorUsr.java
@@ -41,7 +41,7 @@ public class AdaptadorUsr extends BaseAdapter {
 
     @Override
     public long getItemId(int position) {
-        int i= Integer.valueOf(ListaUsuarios.get(position).getId());
+        long i= Long.valueOf(ListaUsuarios.get(position).getId());
         return i;
     }
